+++ packages/eyes-sdk-core/lib/EyesClassic.js
@@ -945,40 +945,6 @@ class EyesClassic extends EyesCore {
     }
   }
   /**
-   * Use this method only if you made a previous call to {@link #open(WebDriver, String, String)} or one of its variants.
-   * @override
-   */
-  async getViewportSize() {
-    const viewportSize = this._viewportSizeHandler.get()
-    return viewportSize
-      ? viewportSize
-      : EyesUtils.getTopContextViewportSize(this._logger, this._driver)
-  }
-  /**
-   * Use this method only if you made a previous call to {@link #open(WebDriver, String, String)} or one of its variants.
-   * @protected
-   * @override
-   */
-  async setViewportSize(viewportSize) {
-    if (this._viewportSizeHandler instanceof ReadOnlyPropertyHandler) {
-      this._logger.verbose('Ignored (viewport size given explicitly)')
-      return Promise.resolve()
-    }
-
-    if (!(await this._controller.isMobile())) {
-      ArgumentGuard.notNull(viewportSize, 'viewportSize')
-      viewportSize = new RectangleSize(viewportSize)
-      try {
-        await EyesUtils.setViewportSize(this._logger, this._driver, new RectangleSize(viewportSize))
-        this._effectiveViewport = new Region(Location.ZERO, viewportSize)
-      } catch (e) {
-        throw new TestFailedError('Failed to set the viewport size', e)
-      }
-    }
-
-    this._viewportSizeHandler.set(new RectangleSize(viewportSize))
-  }
-  /**
    * @override
    */
   async getAppEnvironment() {