+++ packages/eyes-webdriverio-4/test/coverage/index.js
@@ -1,5 +1,4 @@
-const supportedTests = require('./supported-tests')
-const {makeSpecEmitter} = require('@applitools/sdk-coverage-tests')
+const {makeSpecEmitter, supportedTests} = require('@applitools/sdk-coverage-tests')
 
 module.exports = {
   name: 'eyes.webdriverio.javascript4',