+++ packages/sdk-coverage-tests/src/code-export/emit.js
@@ -1,4 +1,4 @@
-const {makeCoverageTests: doMakeCoverageTests} = require('../tests')
+const defaultCoverageTests = require('../tests')
 const {getNameFromObject} = require('../common-util')
 
 function convertExecutionModeToSuffix(executionMode) {
@@ -13,10 +13,18 @@ function convertExecutionModeToSuffix(executionMode) {
   }
 }
 
-function makeEmitTests(initializeSdkImplementation, makeCoverageTests = doMakeCoverageTests) {
+function makeEmitTests(initializeSdkImplementation, coverageTests = defaultCoverageTests) {
   let output = []
   function emitTests(supportedTests, {host, all = false} = {}) {
     supportedTests.forEach(supportedTest => {
+      const coverageTest = coverageTests[supportedTest.name]
+      if (
+        !coverageTest ||
+        !(typeof coverageTest === 'function' || typeof coverageTest.test === 'function')
+      ) {
+        throw new Error('missing implementation for test ' + supportedTest.name)
+      }
+      const coverageTestFunc = coverageTest.test || coverageTest
       const baselineTestName = `${supportedTest.name}${convertExecutionModeToSuffix(
         supportedTest.executionMode,
       )}`
@@ -28,14 +36,10 @@ function makeEmitTests(initializeSdkImplementation, makeCoverageTests = doMakeCo
         branchName,
         host,
         ...supportedTest,
+        ...coverageTest.options,
       })
       // test
-      const coverageTests = makeCoverageTests(sdkImplementation)
-      const coverageTestFunc = coverageTests[supportedTest.name]
-      if (!coverageTestFunc) {
-        throw new Error('missing implementation for test ' + supportedTest.name)
-      }
-      coverageTests[supportedTest.name]()
+      coverageTestFunc(sdkImplementation)
       // store
       output.push({
         name: baselineTestName,
@@ -57,9 +61,45 @@ class EmitTracker {
       afterEach: [],
     }
     this.syntax = {
-      var: null,
+      var: () => {
+        throw new TypeError(
+          "EmitTracker don't have an implementation for `var` syntax. Use `addSyntax` method to add an implementation",
+        )
+      },
+      getter: () => {
+        throw new TypeError(
+          "EmitTracker don't have an implementation for `getter` syntax. Use `addSyntax` method to add an implementation",
+        )
+      },
+      call: () => {
+        throw new TypeError(
+          "EmitTracker don't have an implementation for `call` syntax. Use `addSyntax` method to add an implementation",
+        )
+      },
     }
     this.commands = []
+  }
+
+  createRef(resolve) {
+    const ref = function() {}
+    ref.isRef = true
+    ref.resolve = () => {
+      if (!ref.resolved) {
+        ref.resolved = resolve()
+      }
+      return ref.resolved
+    }
+    return new Proxy(ref, {
+      get: (ref, key) => {
+        if (key in ref) return Reflect.get(ref, key)
+        return this.createRef(() => this.syntax.getter({target: ref.resolve(), key}))
+      },
+      apply: (ref, _, args) => {
+        return this.createRef(() =>
+          this.syntax.call({target: ref.resolve(), args: Array.from(args)}),
+        )
+      },
+    })
   }
 
   addSyntax(name, callback) {
@@ -68,20 +108,12 @@ class EmitTracker {
 
   storeCommand(command) {
     const id = this.commands.push(command)
-    return {
-      isRef: true,
-      resolve: () => {
-        const name = `var_${id}`
-        const value = this.commands[id - 1]
-        if (!this.syntax.var) {
-          throw new TypeError(
-            "EmitTracker don't have an implementation for `var` syntax. Use `addSyntax` method to add an implementation",
-          )
-        }
-        this.commands[id - 1] = this.syntax.var({name, value})
-        return name
-      },
-    }
+    return this.createRef(() => {
+      const name = `var_${id}`
+      const value = this.commands[id - 1]
+      this.commands[id - 1] = this.syntax.var({name, value})
+      return name
+    })
   }
 
   storeHook(name, value) {