+++ packages/sdk-coverage-tests/src/index.js
@@ -1,5 +1,6 @@
 const {makeEmitTracker} = require('./code-export')
 const {makeCoverageTests} = require('./tests')
 const {makeSpecEmitter} = require('./spec-emitter')
+const supportedTests = require('./supported-tests')
 
-module.exports = {makeEmitTracker, makeCoverageTests, makeSpecEmitter}
+module.exports = {makeEmitTracker, makeCoverageTests, makeSpecEmitter, supportedTests}